Fix: Remove `Deck` error handling from `Player` class. * Add code to check if the deck is empty in the `Shiritori.play()` method. * Update the `Deck.isEmpty()` code to check if the `size` of the deck is 0.
